Separated pdf make kits
